Add types for govuk pagination and move govuk types to directory
