--- server/types/govUk.ts
@@ -1,6 +1,0 @@
-type GovUkErrorMessage = {
-  text: string
-}
-
-// eslint-disable-next-line import/prefer-default-export
-export { GovUkErrorMessage }
