+++ server/types/govUk/errorMessage.ts
@@ -1,0 +1,3 @@
+export type GovUkErrorMessage = {
+  text: string
+}
